Fix bug

+++ src/main/java/au/org/aodn/esindexer/model/ExtentModel.java
@@ -3,10 +3,11 @@ package au.org.aodn.esindexer.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @Builder
 public class ExtentModel {
-    protected List<List<Double>> bbox;
+    protected List<List<BigDecimal>> bbox;
 }
